requirements.txt
django==3.2.12
requests==2.9.1
python-decouple==3.6
pymongo==3.5.1